modify the package name.
