+++ src/main/java/io/github/ren2003u/Main.java
@@ -1,0 +1,21 @@
+package io.github.ren2003u;
+
+public class Main {
+    public static void main(String[] args) throws Exception {
+//        UnifiedApiClient client = new UnifiedApiClient("https://ao.space",null);
+//        // Adding a delay to ensure that the APIs are fully updated before proceeding
+//        List<String> serviceIds = Arrays.asList("10001");
+//        String boxUUID = "364b553c01dabb2764b2f2c0e721c1e860e308b1c7daed2671507d21434060ed";
+//        String reqId = "e9993fc787d94b6c886cbaa340f9c0f4";
+//        String sign = "sign";
+//
+//        ObtainBoxRegKeyResponse response = client.obtainBoxRegKey(boxUUID, serviceIds, reqId);
+//        // Blocks until the API response is available
+//        System.out.println("Box UUID: " + response.getBoxUUID());
+//        for (ObtainBoxRegKeyResponse.TokenResult tokenResult : response.getTokenResults()) {
+//            System.out.println("Service Id: " + tokenResult.getServiceId());
+//            System.out.println("Box Reg Key: " + tokenResult.getBoxRegKey());
+//            System.out.println("Expires At: " + tokenResult.getExpiresAt());
+//        }
+    }
+}
