+++ src/main/java/io/github/ren2003u/authentication/model/ObtainBoxRegKeyRequest.java
@@ -1,0 +1,49 @@
+package io.github.ren2003u.authentication.model;
+
+
+import java.util.List;
+
+public class ObtainBoxRegKeyRequest  {
+
+    // Field for Box UUID
+    private String boxUUID;
+
+    // Field for Service IDs
+    private List<String> serviceIds;
+
+    // Field for Sign
+    private String sign;
+
+
+    public String getBoxUUID() {
+        return this.boxUUID;
+    }
+
+    public void setBoxUUID(String boxUUID) {
+        if (boxUUID == null || boxUUID.trim().isEmpty()) {
+            throw new IllegalArgumentException("Box UUID cannot be null or empty.");
+        }
+        this.boxUUID = boxUUID;
+
+    }
+
+    public List<String> getServiceIds() {
+        return this.serviceIds;
+    }
+
+    public void setServiceIds(List<String> serviceIds) {
+        if (serviceIds == null || serviceIds.isEmpty()) {
+            throw new IllegalArgumentException("Service IDs cannot be null or empty.");
+        }
+        this.serviceIds = serviceIds;
+
+    }
+
+    public String getSign() {
+        return this.sign;
+    }
+
+    public void setSign(String sign) {
+        this.sign = sign;
+    }
+}
