+++ src/main/java/io/github/ren2003u/domain/errorHandle/ErrorClass.java
@@ -1,0 +1,46 @@
+package io.github.ren2003u.domain.errorHandle;
+
+public class ErrorClass {
+
+    private int statusCode;
+    private String code;
+    private String message;
+    private String requestId;
+
+
+    public ErrorClass(){
+
+    }
+
+    public int getStatusCode() {
+        return statusCode;
+    }
+
+    public void setStatusCode(int statusCode) {
+        this.statusCode = statusCode;
+    }
+
+    public String getCode() {
+        return code;
+    }
+
+    public void setCode(String code) {
+        this.code = code;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
+
+    public String getRequestId() {
+        return requestId;
+    }
+
+    public void setRequestId(String requestId) {
+        this.requestId = requestId;
+    }
+}
