+++ src/main/java/io/github/ren2003u/register/model/RegisterDeviceRequest.java
@@ -1,0 +1,21 @@
+package io.github.ren2003u.register.model;
+
+
+
+public class RegisterDeviceRequest {
+
+    private String boxUUID;
+
+    public String getBoxUUID() {
+        return this.boxUUID;
+    }
+
+    public void setBoxUUID(String boxUUID) {
+        if (boxUUID == null || boxUUID.trim().isEmpty()) {
+            throw new IllegalArgumentException("Box UUID cannot be null or empty.");
+        }
+        this.boxUUID = boxUUID;
+
+    }
+
+}
