--- src/main/java/org/example/domain/errorHandle/ErrorClass.java
@@ -1,46 +1,0 @@
-package org.example.domain.errorHandle;
-
-public class ErrorClass {
-
-    private int statusCode;
-    private String code;
-    private String message;
-    private String requestId;
-
-
-    public ErrorClass(){
-
-    }
-
-    public int getStatusCode() {
-        return statusCode;
-    }
-
-    public void setStatusCode(int statusCode) {
-        this.statusCode = statusCode;
-    }
-
-    public String getCode() {
-        return code;
-    }
-
-    public void setCode(String code) {
-        this.code = code;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-
-    public String getRequestId() {
-        return requestId;
-    }
-
-    public void setRequestId(String requestId) {
-        this.requestId = requestId;
-    }
-}
