--- src/main/java/org/example/register/model/RegisterDeviceRequest.java
@@ -1,21 +1,0 @@
-package org.example.register.model;
-
-
-
-public class RegisterDeviceRequest {
-
-    private String boxUUID;
-
-    public String getBoxUUID() {
-        return this.boxUUID;
-    }
-
-    public void setBoxUUID(String boxUUID) {
-        if (boxUUID == null || boxUUID.trim().isEmpty()) {
-            throw new IllegalArgumentException("Box UUID cannot be null or empty.");
-        }
-        this.boxUUID = boxUUID;
-
-    }
-
-}
