+++ src/main/resources/logback.xml
@@ -6,7 +6,7 @@
         </encoder>
     </appender>
 
-    <logger name="org.example.client.Client" level="INFO">
+    <logger name="io.github.ren2003u.client.Client" level="INFO">
         <appender-ref ref="FILE" />
     </logger>
